Vijeth Dinesha's bug fix, compile on mac
Ignore-this: b9fee3bfa39a4b33348b822408ff31db

darcs-hash:20100519204315-889e6-b2a435b1dba0dcbb660b765f813582a2f0ee66e3

+++ examples/trilinear/main.c
@@ -293,7 +293,7 @@ main (int argc__, char *argv__[])
     }
 
     size_t size = sizei[0]*sizei[1]*sizei[2];
-    size_t sizeu[3] = {sizei[0],sizei[1],sizei[2]};
+    uint32_t sizeu[3] = {sizei[0],sizei[1],sizei[2]};
    
     struct stat s;
     stat(argv[4],&s);
